Factor out VotingRoundParticipants component.
